Changing to direct threading model
Changed the interpreter to direct threading model. It is fairly straightforward: native words don't need any prologue anymore, and Forth words use "jmp call" instead of "!word call". The interpreter core is modified and some words (particularly, creation words) are fixed to work with the new model. As there are more Forth words than native words, the interpreter is a little bit bigger, but the difference is negligible. On the positive side, it is about 10% faster. This change will make it easier to experiment with switching to token threaded model (otherwise it would require double dereferencing). I may also remove the hack of moving the interpreter to zero page, as it is no longer needed.