Updates in progress
Rather significant batch of updates. The binary size is down from 11K to 9.7K. Besides a few minor non-standard things removed, the biggest effects were achieved by changing Forth word prolog to just RTS (CALL sets up the stack so RTS acts up as JMP) - that was more than 400 bytes - and changing backward word link to a variable length offset saving another 300 bytes. Minimal effect on performance, much less than expected. The code is a bit ugly, but slowly cleaning up bits and pieces.
Now, ready for the last push toward the 8K size. It actually looks achievable. The idea is to switch to token threaded model and ensure that all tokens used in the core are one byte in size. Considering that there are less than 200 unique tokens used in the core and ~2,300 tokens used total, it may be possible to save more than 2K with this change alone. It will require some code overhead and runtime token table generation, but there is some slack for that. On top of that, there are a few places where some bytes can still be saved.