Updates and fixes
Another round of updates - this version can run the test suite unmodified! The interpreter is case-insensitive once again. Due to peculiarities of PETSCII, it is not as clean as it would be on other systems - to make the test output legible the system is switched to lowercase mode (use "142 EMIT" to switch back once done). It looks a bit odd, but works fine.
Data stack moved together with the rest of high memory data, away from $0400. This decreases available RAM by 1K, but that will be recovered later when memory models get finalized - on C64 there is a 4K block of RAM available at $C000, pretty sure Commander X16 can spare some of its memory as well.
The core runs on C64, but the file I/O is not working properly. Not a big surprise, but it is the next thing to address. It is not clear if it works on X16 outside of emulator either.